Remove boot and modules volumes
Qemu artifacts are handled via libguestfs which need a kernel.
By default it uses what it found in /boot but docker have not ones and
so historicaly, lava-docker created a /boot volume.
But this is bad, on host with custom kernel (like mine), kernel could miss
some necessary VIRTIO drivers.
Furthermore, this is bad to give access to host kernel from a security point of view.

So remove volumes and install a debian kernel on slave.
For master, in fact it needed nothing.

Signed-off-by: Corentin LABBE <clabbe@baylibre.com>